Add PublicAPI and UsedImplicity attributes where applicable
