+++ source/Nuke.Common/CI/AppVeyor/Configuration/AppVeyorBranches.cs
@@ -4,11 +4,13 @@
 
 using System;
 using System.Linq;
+using JetBrains.Annotations;
 using Nuke.Common.Utilities;
 using Nuke.Common.Utilities.Collections;
 
 namespace Nuke.Common.CI.AppVeyor.Configuration
 {
+    [PublicAPI]
     public class AppVeyorBranches : ConfigurationEntity
     {
         public string[] Only { get; set; }
