+++ source/Nuke.Common/CI/AzurePipelines/Configuration/AzurePipelinesStage.cs
@@ -1,11 +1,13 @@
 using System;
 using System.Linq;
+using JetBrains.Annotations;
 using Nuke.Common.Tooling;
 using Nuke.Common.Utilities;
 using Nuke.Common.Utilities.Collections;
 
 namespace Nuke.Common.CI.AzurePipelines.Configuration
 {
+    [PublicAPI]
     public class AzurePipelinesStage : ConfigurationEntity
     {
         public string Name { get; set; }
