+++ source/Nuke.Common/CI/GitHubActions/Configuration/GitHubActionsDetailedTrigger.cs
@@ -4,9 +4,11 @@
 
 using System;
 using System.Linq;
+using JetBrains.Annotations;
 
 namespace Nuke.Common.CI.GitHubActions.Configuration
 {
+    [PublicAPI]
     public abstract class GitHubActionsDetailedTrigger : ConfigurationEntity
     {
     }
