+++ source/Nuke.Common/CI/GitHubActions/Configuration/GitHubActionsUsingStep.cs
@@ -4,11 +4,12 @@
 
 using System;
 using System.Linq;
+using JetBrains.Annotations;
 using Nuke.Common.Utilities;
-using Nuke.Common.Utilities.Collections;
 
 namespace Nuke.Common.CI.GitHubActions.Configuration
 {
+    [PublicAPI]
     public class GitHubActionsUsingStep : GitHubActionsStep
     {
         public string Using { get; set; }
@@ -18,29 +19,7 @@ namespace Nuke.Common.CI.GitHubActions.Configuration
             writer.WriteLine($"- uses: {Using}");
         }
     }
-    public class GitHubActionsCacheStep : GitHubActionsStep
-    {
-        public string[] IncludePatterns { get; set; }
-        public string[] ExcludePatterns { get; set; }
-        public string[] KeyFiles { get; set; }
 
-        public override void Write(CustomFileWriter writer)
-        {
-            writer.WriteLine($"- name: Cache {IncludePatterns.JoinComma()}");
-            using (writer.Indent())
-            {
-                writer.WriteLine("uses: actions/cache@v2");
-                writer.WriteLine("with:");
-                using (writer.Indent())
-                {
-                    writer.WriteLine("path: |");
-                    IncludePatterns.ForEach(x => writer.WriteLine($"  {x}"));
-                    ExcludePatterns.ForEach(x => writer.WriteLine($"  !{x}"));
-                    writer.WriteLine($"key: ${{{{ runner.os }}}}-${{{{ hashFiles({KeyFiles.Select(x => x.SingleQuote()).JoinComma()}) }}}}");
-                }
-            }
-        }
-    }
     public class GitHubActionsArtifactStep : GitHubActionsStep
     {
         public string Name { get; set; }
