+++ source/Nuke.Common/CI/SpaceAutomation/Configuration/SpaceAutomationConfiguration.cs
@@ -3,11 +3,13 @@
 // https://github.com/nuke-build/nuke/blob/master/LICENSE
 
 using System.Linq;
+using JetBrains.Annotations;
 using Nuke.Common.Utilities;
 using Nuke.Common.Utilities.Collections;
 
 namespace Nuke.Common.CI.SpaceAutomation.Configuration
 {
+    [PublicAPI]
     public class SpaceAutomationConfiguration : ConfigurationEntity
     {
         public string Name { get; set; }
