+++ source/Nuke.Common/CI/SpaceAutomation/Configuration/SpaceAutomationContainer.cs
@@ -4,10 +4,12 @@
 
 using System;
 using System.Linq;
+using JetBrains.Annotations;
 using Nuke.Common.Utilities;
 
 namespace Nuke.Common.CI.SpaceAutomation.Configuration
 {
+    [PublicAPI]
     public class SpaceAutomationContainer : ConfigurationEntity
     {
         public string Image { get; set; }
