+++ source/Nuke.Common/CI/SpaceAutomation/Configuration/SpaceAutomationTrigger.cs
@@ -1,8 +1,10 @@
 ﻿using System;
 using System.Linq;
+using JetBrains.Annotations;
 
 namespace Nuke.Common.CI.SpaceAutomation.Configuration
 {
+    [PublicAPI]
     public abstract class SpaceAutomationTrigger : ConfigurationEntity
     {
     }
