+++ source/Nuke.Common/CI/TeamCity/Configuration/TeamCityConfigurationParameter.cs
@@ -5,10 +5,12 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using JetBrains.Annotations;
 using Nuke.Common.Utilities;
 
 namespace Nuke.Common.CI.TeamCity.Configuration
 {
+    [PublicAPI]
     public class TeamCityConfigurationParameter : TeamCityParameter
     {
         public TeamCityParameterType Type { get; set; }
