+++ source/Nuke.Common/CI/TeamCity/Configuration/TeamCityParameter.cs
@@ -4,9 +4,11 @@
 
 using System;
 using System.Linq;
+using JetBrains.Annotations;
 
 namespace Nuke.Common.CI.TeamCity.Configuration
 {
+    [PublicAPI]
     public abstract class TeamCityParameter : ConfigurationEntity
     {
     }
