+++ source/Nuke.Common/Execution/BuildExtensionAttributeBase.cs
@@ -5,25 +5,30 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using JetBrains.Annotations;
 
 namespace Nuke.Common.Execution
 {
+    [PublicAPI]
     public interface IBuildExtension
     {
         float Priority { get; }
     }
 
+    [PublicAPI]
     [AttributeUsage(AttributeTargets.Class)]
     public abstract class BuildExtensionAttributeBase : Attribute, IBuildExtension
     {
         public virtual float Priority { get; set; }
     }
 
+    [PublicAPI]
     public interface IOnBeforeLogo : IBuildExtension
     {
         void OnBeforeLogo(NukeBuild build, IReadOnlyCollection<ExecutableTarget> executableTargets);
     }
 
+    [PublicAPI]
     public interface IOnAfterLogo : IBuildExtension
     {
         void OnAfterLogo(
@@ -32,26 +37,31 @@ namespace Nuke.Common.Execution
             IReadOnlyCollection<ExecutableTarget> executionPlan);
     }
 
+    [PublicAPI]
     public interface IOnTargetSkipped : IBuildExtension
     {
         void OnTargetSkipped(NukeBuild build, ExecutableTarget target);
     }
 
+    [PublicAPI]
     public interface IOnTargetStart : IBuildExtension
     {
         void OnTargetStart(NukeBuild build, ExecutableTarget target);
     }
 
+    [PublicAPI]
     public interface IOnTargetExecuted : IBuildExtension
     {
         void OnTargetExecuted(NukeBuild build, ExecutableTarget target);
     }
 
+    [PublicAPI]
     public interface IOnTargetFailed : IBuildExtension
     {
         void OnTargetFailed(NukeBuild build, ExecutableTarget target);
     }
 
+    [PublicAPI]
     public interface IOnBuildFinished : IBuildExtension
     {
         void OnBuildFinished(NukeBuild build);
