+++ source/Nuke.Common/Tooling/ExtensionHelper.cs
@@ -6,11 +6,13 @@ using System;
 using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
+using JetBrains.Annotations;
 using Nuke.Common.Utilities;
 using static Nuke.Common.Utilities.ReflectionUtility;
 
 namespace Nuke.Common.Tooling
 {
+    [UsedImplicitly(ImplicitUseTargetFlags.WithMembers)]
     public static class ExtensionHelper
     {
         public static void ToggleBoolean(IDictionary dictionary, string key)
